requirements.txt
python-fasthtml>=0.6.0
python-dotenv>=1.0
gunicorn>=21.0
uvicorn>=0.30.0
